compute can do perfect_num finding jobs now

+++ lab4/defs.h
@@ -9,9 +9,9 @@
 #include <sys/types.h>
 
 #define KEY (key_t)78738
-#define BMAP_SIZE (size_t)(1 << 22)
+#define BMAP_SIZE (1 << 22)
 #define MAX_PRC_NUM 20
-#define SHM_SIZE BMAP_SIZE + MAX_PRC_NUM * sizeof(int) + MAX_PRC_NUM * sizeof(process_t) + sizeof(pid_t) + sizeof(process_t)
+#define SHM_SIZE (size_t)BMAP_SIZE + MAX_PRC_NUM * sizeof(int) + MAX_PRC_NUM * sizeof(process_t) + sizeof(pid_t) + sizeof(process_t)
 #define MSG_TYPE_REGISTER 1
 #define MSG_TYPE_PREPORT 1
